pyproject.toml
[project]
name = "langchain-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.14"
dependencies = [
    "fastapi[standard]>=0.135.1",
    "langchain-ollama>=1.0.1",
    "ollama>=0.6.1",
    "python-dotenv>=1.2.2",
    "streamlit>=1.55.0",
]
